pyvideotrans_cli/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pyvideotrans-cli"
version = "0.1.0"
description = "精简版视频翻译 CLI 工具 - 支持 Whisper 转录、本地 LLM 翻译、Qwen TTS 配音"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Customized from pyvideotrans"}
]
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "faster-whisper>=1.0.0",
    "openai-whisper>=20230314",
    "openai>=1.0.0",
    "httpx>=0.25.0",
    "dashscope>=1.14.0",
    "requests>=2.31.0",
    "pydub>=0.25.1",
    "ffmpeg-python>=0.2.0",
    "transformers>=4.40.0",
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    "accelerate>=0.25.0",
    "soundfile>=0.12.1",
]

[project.optional-dependencies]
cuda = [
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
]
qwen-local = [
    "transformers>=4.40.0",
    "accelerate>=0.25.0",
    "soundfile>=0.12.1",
]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
pyvideotrans-cli = "pyvideotrans_cli.cli:main"

[project.urls]
Homepage = "https://github.com/jianchang512/pyvideotrans"
Repository = "https://github.com/jianchang512/pyvideotrans.git"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"*" = ["*.json", "*.txt"]

[tool.black]
line-length = 100
target-version = ["py310", "py311"]

[tool.ruff]
line-length = 100
target-version = "py310"
